# Conflicts:
#	Source/Lab1/OptimisationMethodRunnercs.cs
#	Source/Lab1/OptimisationMethods/DichotomyMethod.cs
#	Source/Lab1/OptimisationMethods/FibonacciMethod.cs
#	Source/Lab1/OptimisationMethods/GoldenRatioMethod.cs
#	Source/Lab1/OptimisationMethods/IOptimisationMethod.cs
#	Source/Lab1/OptimisationMethods/ParabolaMethod.cs
#	Source/Lab1/Program.cs
#	Source/Lab1/RunnerResult.cs

+++ Source/Lab1/OptimisationMethods/FibonacciMethod.cs
@@ -6,15 +6,15 @@ namespace Lab1.OptimisationMethods;
 public class FibonacciMethod : IOptimisationMethod<BoundedOptimizationContext>
 {
     private readonly int _n;
-    private decimal? _x1;
-    private decimal? _x2;
+    private double? _x1;
+    private double? _x2;
 
     public FibonacciMethod(int n)
     {
         _n = n;
     }
 
-    public BoundedOptimizationContext FindNewInterval(BoundedOptimizationContext context, Func<decimal, decimal> function)
+    public BoundedOptimizationContext FindNewInterval(BoundedOptimizationContext context, Func<double, double> function)
     {
         var (a, b) = (context.A, context.B);
         var x1 = _x1 ?? a + (FibonacciCounter.GetNthNumber(_n - 2) / FibonacciCounter.GetNthNumber(_n)) * (b-a);
