+++ Source/Lab1/Tools/FibonacciCounter.cs
@@ -2,9 +2,9 @@
 
 public static class FibonacciCounter
 {
-    private static List<decimal> _cache = new List<decimal>(){1, 1};
+    private static List<double> _cache = new List<double>(){1, 1};
     
-    public static decimal GetNthNumber(int n)
+    public static double GetNthNumber(int n)
     {
         if (n < 0)
             throw new ArgumentException("N must be positive or 0");
